CloudQSim.jl

Documentation for CloudQSim.jl will be here soon. Meanwhile, out the github repo.

CloudQSim — Module

Cloud simulation of quantum evolution using Bloqade.jl

Module Index

CloudQSim.CloudConfig

CloudQSim.CloudQSimTask

CloudQSim.cloud_simulate

CloudQSim.get_last_meta

CloudQSim.split_task

Detailed API

CloudQSim.CloudConfig — Type

Configuration for cloud simulation servers.

Fields

addrs: Hosts to connect to

ports: Ports to use

worker_counts: Number of workers on each server to use in load balancing

Example

julia> clconf = CloudConfig()
julia> add_server!(clconf, "cloudqs.lykov.tech", 7700)

CloudQSim.CloudQSimTask — Type

Simple wrapper for cloud server task

CloudQSim.cloud_simulate — Method

Run quantum evolution of multiple Hamiltonians in the cloud.

Arguments

hamiltonian::Vector: List of Hamiltonians

time_points::Int: How many evaluations of observables to do throughout the simulation

observables::Vector{<:Vector}: List of observables to evaluate

subspace_radius::Float64=0.: Radius of the subspace to use for the Bloqade subspace evolution

Returns

results::Array: Array of results, with dimensions (hamiltonians, time_points, observables)

CloudQSim.get_last_meta — Method

Get metadata about the last simulation, such as time spent in network, parsing results, etc.

CloudQSim.split_task — Method

Split one CloudQSimTask into many for balancing of jobs betwen worker servers