CloudQSim.jl

Documentation for CloudQSim.jl package

Usage

To start simulating on the cloud, you will need:

Hostname and port of the server. cloudqs.lykov.tech:7700 is the demo server

CLODUQS_TOKEN - authorization token to run tasks

A Bloqade.jl hamiltonian

A set of observables to measure. See observables

Please refer to cloud_simulate function for the main API.

CloudQSim — Module

Cloud simulation of quantum evolution using Bloqade.jl

Module Index

CloudQSim.CloudConfig

CloudQSim.CloudQSimTask

CloudQSim.cloud_simulate

CloudQSim.get_last_meta

CloudQSim.split_task

Creating observables for simulation

The default format for observables is a diagonal observable, which is represented by a vector where each element represents the observable value for each corresponding state.

While a general observable may be to expensive to send, most non-diagonal observables can be represented as a sum of local terms.

The following definition of observables describes how to specify such set of observables.

Using observables

The observable expression is defined as a sum of local terms applied to different qubits.

Create definitions of local terms

Apply the local terms to qubits and add to the sum

Note that one can use any local observable as defined by matrix or $n$-dimensional array.

Example:

obs = CloudQSObservable()
X = define_operator!(obs, [[0, 1], [1, 0]], nqubits=1)
XX = define_operator!(obs, ..., nqubits=2)

set_sum!(obs, X[0], X[1], X[2], XX[0, 2], XX[0, 1])

res = cloud_simulate(hams, obs, clconf)

Currently, the only way to define a local operator is to provide the full matrix. In other words, one has to obtain the matrix by using for example, Yao:

using Yao
XX_mat = mat(kron(X for _ in 1:2))

It future, this package may provide a more native way to construct observables in a way QSpin does it.

Data structure

This is the string that will be sent over to the server:

{
    defs: {
        "X": ndarray
        "XX": ndarray
        },
    sum_terms: [
        {label: "X", qubits: [0]},
        {label: "X", qubits: [1]},
        {label: "X", qubits: [2]},
        {label: "XX", qubits: [0, 2]},
        {label: "XX", qubits: [0, 1]},
    ]
}

Detailed API

CloudQSim.CloudConfig — Type

Configuration for cloud simulation servers.

Fields

addrs: Hosts to connect to

ports: Ports to use

worker_counts: Number of workers on each server to use in load balancing

Example

julia> clconf = CloudConfig()
julia> add_server!(clconf, "cloudqs.lykov.tech", 7700)

CloudQSim.CloudQSimTask — Type

Simple wrapper for cloud server task

CloudQSim.cloud_simulate — Method

Run quantum evolution of multiple Hamiltonians in the cloud.

Arguments

hamiltonian::Vector: List of Hamiltonians

time_points::Int: How many evaluations of observables to do throughout the simulation

observables::Vector{<:Vector}: List of observables to evaluate

subspace_radius::Float64=0.: Radius of the subspace to use for the Bloqade subspace evolution

Returns

results::Array: Array of results, with dimensions (hamiltonians, time_points, observables)

CloudQSim.get_last_meta — Method

Get metadata about the last simulation, such as time spent in network, parsing results, etc.

CloudQSim.split_task — Method

Split one CloudQSimTask into many for balancing of jobs betwen worker servers